+++ src/Rede/Rede.java
@@ -304,9 +304,9 @@ public class Rede {
     private void calculaErroDaRede() {
         erro = 0f;
         for(NeuronioSaida neuronio:listaNeuronioSaida){
-            erro += Math.pow(neuronio.getErro(),2);
+            erro += Math.pow(neuronio.getErro(),(float)2);
         }
-        erro /= 2;
+        erro /= (float)2;
     }
 
     private void calculaNETCamadaOculta() {
